pipeline/assets/requirements.txt
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
pandas>=2.0.0
# add other dependencies your pipeline uses below
